✨ [feat] 실시간/파리 상품 섹션 분리
